ocular/providers/local.py
from transformers import pipeline, AutoTokenizer, AutoModelForCausalLM
import torch
import json

class LocalLLMParser:
    def __init__(self, model_name: str = "microsoft/DialoGPT-medium"):
        self.tokenizer = AutoTokenizer.from_pretrained(model_name)
        self.model = AutoModelForCausalLM.from_pretrained(model_name)
        
        # Add padding token if it doesn't exist
        if self.tokenizer.pad_token is None:
            self.tokenizer.pad_token = self.tokenizer.eos_token
    
    def parse_to_json(self, unstructured_text: str) -> Dict[str, Any]:
        """Convert unstructured text to JSON using local model"""
        
        prompt = f"""
        Convert this text to JSON:
        Text: {unstructured_text}
        JSON:
        """
        
        # This is a simplified example - you'd need a model trained for this task
        inputs = self.tokenizer.encode(prompt, return_tensors="pt", truncation=True, max_length=512)
        
        with torch.no_grad():
            outputs = self.model.generate(
                inputs,
                max_length=inputs.shape[1] + 200,
                temperature=0.7,
                pad_token_id=self.tokenizer.pad_token_id
            )
        
        response = self.tokenizer.decode(outputs[0], skip_special_tokens=True)
        # Extract JSON from response (this would need more sophisticated parsing)
        
        return {"note": "Local models need fine-tuning for JSON extraction"}